Make Handler::before() hierarchy consistent
In API, the before() method (inherited from Handler) takes a $method
argument, although this is not declared for the before method() in
handler.  PHP in strict mode treats this as an error.

Since call sites already pass the $method argument, change
Handler::before() and the before() method in all other subclasses to
take this argument.