Added maps and nl_NL translation files
sorry, made a mistake by creating a file instead of map. So that one
has to be deleted.  Added in the right way the necessary folders and
added the messages.mo and po files for the Dutch translation. I suppose
you yourself will changes the rest to include Dutch. I have done that
for my own version (added nl_NL to the function.php file).
regards, GvMelle