document the date format string
I assumed the format string was related to strftime except all of the
fields looked really weird.  After digging in the code, this string is
actually used with php's date().  Add a note for this so other people
can jump quickly to the relevant documentation (which is useful for
non-php peeps who have no idea what this is).